skip tileserver startup when mbtiles not found

+++ startup.sh
@@ -17,9 +17,12 @@ if test -f "$TILESERVER_PID"; then
     pkill -F $TILESERVER_PID
 fi
 
-echo "run tileserver"
-tileserver-gl-light ./extract.mbtiles &
-echo $! > $TILESERVER_PID
+TILE_PATH=./extract.mbtiles
+if test -f "$TILE_PATH"; then
+    echo "run tileserver"
+    tileserver-gl-light $TILE_PATH &
+    echo $! > $TILESERVER_PID
+fi
 
 sudo modprobe ledtrig_heartbeat
 sudo su root -c 'echo heartbeat >/sys/class/leds/led0/trigger'
